Add Analytics API and refactor package structure to separate telemetry ingestion from analytics concerns
