--- src/main/java/com/fahrisnassief/dronetelemetry/DroneIdDTO.java
@@ -1,3 +1,0 @@
-package com.fahrisnassief.dronetelemetry;
-
-public record DroneIdDTO(String droneID) {}
