+++ src/main/java/com/fahrisnassief/dronetelemetry/analytics/AnalyticsRepository.java
@@ -1,11 +1,12 @@
-package com.fahrisnassief.dronetelemetry;
+package com.fahrisnassief.dronetelemetry.analytics;
 
+import com.fahrisnassief.dronetelemetry.model.DroneTelemetry;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface DroneTelemetryRepository extends JpaRepository <DroneTelemetry, Long> {
+public interface AnalyticsRepository extends JpaRepository<DroneTelemetry, Long> {
 
     @Query("SELECT DISTINCT d.droneId FROM DroneTelemetry d")
     List<String> findDistinctDrones();
