+++ src/main/java/com/fahrisnassief/dronetelemetry/analytics/DroneIdDTO.java
@@ -1,0 +1,3 @@
+package com.fahrisnassief.dronetelemetry.analytics;
+
+public record DroneIdDTO(String droneID) {}
