+++ src/main/java/com/fahrisnassief/dronetelemetry/model/DroneTelemetry.java
@@ -1,4 +1,4 @@
-package com.fahrisnassief.dronetelemetry;
+package com.fahrisnassief.dronetelemetry.model;
 
 import jakarta.persistence.*;
