+++ src/main/java/com/fahrisnassief/dronetelemetry/telemetry/DroneTelemetryRepository.java
@@ -1,0 +1,6 @@
+package com.fahrisnassief.dronetelemetry.telemetry;
+
+import com.fahrisnassief.dronetelemetry.model.DroneTelemetry;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface DroneTelemetryRepository extends JpaRepository <DroneTelemetry, Long> {}
