+++ src/main/java/com/fahrisnassief/dronetelemetry/telemetry/DroneTelemetryService.java
@@ -1,0 +1,13 @@
+package com.fahrisnassief.dronetelemetry.telemetry;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class DroneTelemetryService {
+
+    private final DroneTelemetryRepository droneTelemetryRepository;
+
+    public DroneTelemetryService(DroneTelemetryRepository droneTelemetryRepository) {
+        this.droneTelemetryRepository = droneTelemetryRepository;
+    }
+}
